// src/app/types/enum/code-type.enum.ts
export enum CodeType {
  None = 0,
  Ansi39 = 1,
  Ansi39x = 2,
  Code_2_of_5 = 3,
  Code25intlv = 4,
  Matrix_2_of_5 = 5,
  Code39 = 6,
  Code39x = 7,
  Code_128_A = 8,
  Code_128_B = 9,
  Code_128_C = 10,
  Code_128auto = 11,
  Code_93 = 12,
  Code93x = 13,
  MSI = 14,
  PostNet = 0xf,
  Codabar = 0x10,
  EAN_8 = 17,
  EAN_13 = 18,
  UPC_A = 19,
  UPC_E0 = 20,
  UPC_E1 = 21,
  RM4SCC = 22,
  UCCEAN128 = 23,
  QRCode = 24,
  Code49 = 25,
  JapanesePostal = 26,
  Pdf417 = 27,
  EAN128FNC1 = 28,
  RSS14 = 29,
  RSS14Truncated = 30,
  RSS14Stacked = 0x1f,
  RSS14StackedOmnidirectional = 0x20,
  RSSExpanded = 33,
  RSSExpandedStacked = 34,
  RSSLimited = 35,
  DataMatrix = 36,
  MicroPDF417 = 37,
  IntelligentMail = 0x40,
}